Changed vertex format in video memory (packed in 4bytes).

+++ source/core/render/opengl_chunk.cpp
@@ -3,7 +3,7 @@
 
 namespace blocks
 {
-  OpenglChunk::OpenglChunk(std::shared_ptr<OpenglBuffer> vbo, std::shared_ptr<OpenglVertexArrayObject> vao, int verticesNumber)
+  OpenglChunk::OpenglChunk(std::shared_ptr<OpenglBuffer> vbo, std::shared_ptr<OpenglVertexArrayObject> vao, size_t verticesNumber)
     : vbo_(vbo), vao_(vao), verticesNumber_(verticesNumber)
   {
